requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
attrs==23.2.0
Faker==23.2.0
frozenlist==1.4.1
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
peewee==3.17.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
yarl==1.9.4
